Split pr_cmd_cards and large runtime modules
